backend/api-gateway/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
httpx==0.24.0
python-dotenv==1.0.0
python-jose==3.3.0
pydantic==1.10.7
PyJWT==2.6.0
requests==2.29.0
